requirements.txt
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
asyncpg==0.31.0
attrs==25.4.0
backcall==0.2.0
beautifulsoup4==4.14.3
bleach==6.3.0
blue==0.9.1
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.4
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
dnspython==2.8.0
docopt==0.6.2
ecdsa==0.19.1
email-validator==2.3.0
executing==2.2.1
fastapi==0.124.0
fastjsonschema==2.21.2
fsspec==2026.1.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iso8601==2.1.0
isort==7.0.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mccabe==0.6.1
mdurl==0.1.2
mistune==3.2.0
mmh3==5.2.0
multidict==6.7.1
mypy_extensions==1.1.0
nbclient==0.10.4
nbformat==5.10.4
packaging==26.0
pandocfilters==1.5.1
parso==0.8.5
passlib==1.7.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.5.1
postgrest==2.27.2
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2==2.9.11
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pycodestyle==2.8.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==2.4.0
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pyparsing==3.3.2
pypika-tortoise==0.6.3
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
pyzmq==27.1.0
realtime==2.27.2
referencing==0.37.0
requests==2.32.5
rich==14.3.1
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
sortedcontainers==2.4.0
soupsieve==2.8.3
stack-data==0.6.3
starlette==0.50.0
storage3==2.27.2
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.2
supabase-auth==2.27.2
supabase-functions==2.27.2
tenacity==9.1.2
tinycss2==1.4.0
tomli==2.3.0
tornado==6.5.4
tortoise-orm==0.25.1
traitlets==5.14.3
typing_extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.38.0
wcwidth==0.5.3
webencodings==0.5.1
websockets==15.0.1
yarg==0.1.9
yarl==1.22.0
bcrypt
